Importer Uploading Pane: improve keyboard accessibility
This patch not only removes all ESLint jsx-a11y warnings, but actually improves the
keyboard accessibility, too. The Uploader Pane reacts to Enter and Space keys by opening
the file selection dialog, and changes border color to a different shade of gray when
it gets focus. That can be tested by tabbing to the pane.